tests - dont allow to run against prod, maintainers cleanup fixes

+++ tests/ClientTestCase.php
@@ -13,6 +13,16 @@ use Keboola\ManageApi\Client;
 class ClientTestCase extends \PHPUnit_Framework_TestCase
 {
 
+    const PRODUCTION_HOSTS = [
+        'connection.keboola.com',
+        'connection.eu-central-1.keboola.com',
+    ];
+
+    /**
+     * Prefix of all maintainers created by tests
+     */
+    const TESTS_MAINTAINER_PREFIX = 'KBC_MANAGE_TESTS';
+
     /**
      * @var Client
      */
@@ -31,9 +41,15 @@ class ClientTestCase extends \PHPUnit_Framework_TestCase
 
     public static function setUpBeforeClass()
     {
+        $manageApiUrl = getenv('KBC_MANAGE_API_URL');
+
+        if (in_array(parse_url($manageApiUrl, PHP_URL_HOST), self::PRODUCTION_HOSTS)) {
+            throw new \Exception('Tests cannot be executed against production host - ' . $manageApiUrl);
+        }
+
         $client = new Client([
             'token' => getenv('KBC_MANAGE_API_TOKEN'),
-            'url' => getenv('KBC_MANAGE_API_URL'),
+            'url' => $manageApiUrl,
             'backoffMaxTries' => 0,
         ]);
         $organizations = $client->listMaintainerOrganizations(getenv('KBC_TEST_MAINTAINER_ID'));
@@ -61,6 +77,8 @@ class ClientTestCase extends \PHPUnit_Framework_TestCase
 
         $this->normalUser = $this->normalUserClient->verifyToken()['user'];
         $this->superAdmin = $this->client->verifyToken()['user'];
+
+        // cleanup maintainers created by tests
         $maintainers = $this->client->listMaintainers();
         foreach ($maintainers as $maintainer) {
             if ((int) $maintainer['id'] === (int) $this->testMaintainerId) {
@@ -70,7 +88,7 @@ class ClientTestCase extends \PHPUnit_Framework_TestCase
                         $this->client->removeUserFromMaintainer($maintainer['id'], $member['id']);
                     }
                 }
-            } else {
+            } elseif (strpos($maintainer['name'], self::TESTS_MAINTAINER_PREFIX) === 0) {
                 $this->client->deleteMaintainer($maintainer['id']);
             }
         }
